--- a/cpp/include/cuml/ensemble/randomforest.hpp
+++ b/cpp/include/cuml/ensemble/randomforest.hpp
@@ -188,11 +188,12 @@ RF_metrics score(const raft::handle_t& user_handle,
 
 RF_params set_rf_class_obj(int max_depth, int max_leaves, float max_features,
                            int n_bins, int split_algo, int min_rows_per_node,
-                           float min_impurity_decrease, bool bootstrap_features,
-                           bool bootstrap, int n_trees, float rows_sample,
-                           int seed, CRITERION split_criterion,
-                           bool quantile_per_tree, int cfg_n_streams,
-                           bool use_experimental_backend, int max_batch_size);
+                           int min_samples_split, float min_impurity_decrease,
+                           bool bootstrap_features, bool bootstrap, int n_trees,
+                           float rows_sample, int seed,
+                           CRITERION split_criterion, bool quantile_per_tree,
+                           int cfg_n_streams, bool use_experimental_backend,
+                           int max_batch_size);
 
 // ----------------------------- Regression ----------------------------------- //