--- a/cpp/include/cuml/tree/decisiontree.hpp
+++ b/cpp/include/cuml/tree/decisiontree.hpp
@@ -50,6 +50,10 @@ struct DecisionTreeParams {
    */
   int min_rows_per_node;
   /**
+   * The minimum number of samples (rows) needed to split an internal node.
+   */
+  int min_samples_split;
+  /**
    * Control bootstrapping for features. If features are drawn with or without replacement
    */
   bool bootstrap_features;
@@ -91,6 +95,8 @@ struct DecisionTreeParams {
  * @param[in] cfg_n_bins: number of bins; default 8
  * @param[in] cfg_split_algo: split algorithm; default SPLIT_ALGO::HIST
  * @param[in] cfg_min_rows_per_node: min. rows per node; default 2
+ * @param[in] cfg_min_samples_split: min. rows needed to split an internal node;
+ *            default 2
  * @param[in] cfg_min_impurity_decrease: split a node only if its reduction in
  *                                       impurity is more than this value
  * @param[in] cfg_bootstrap_features: bootstrapping for features; default false
@@ -108,6 +114,7 @@ void set_tree_params(DecisionTreeParams &params, int cfg_max_depth = -1,
                      int cfg_max_leaves = -1, float cfg_max_features = 1.0f,
                      int cfg_n_bins = 8, int cfg_split_algo = SPLIT_ALGO::HIST,
                      int cfg_min_rows_per_node = 2,
+                     int cfg_min_samples_split = 2,
                      float cfg_min_impurity_decrease = 0.0f,
                      bool cfg_bootstrap_features = false,
                      CRITERION cfg_split_criterion = CRITERION_END,