--- a/cpp/src/decisiontree/batched-levelalgo/builder_base.cuh
+++ b/cpp/src/decisiontree/batched-levelalgo/builder_base.cuh
@@ -401,7 +401,7 @@ struct ClsTraits {
     CUDA_CHECK(cudaMemsetAsync(b.hist, 0, sizeof(int) * b.nHistBins, s));
     computeSplitClassificationKernel<DataT, LabelT, IdxT, TPB_DEFAULT>
       <<<grid, TPB_DEFAULT, smemSize, s>>>(
-        b.hist, b.params.n_bins, b.params.max_depth, b.params.min_samples_leaf,
+        b.hist, b.params.n_bins, b.params.max_depth, b.params.min_samples_split,
         b.params.max_leaves, b.input, b.curr_nodes, col, b.done_count, b.mutex,
         b.n_leaves, b.splits, splitType);
   }
@@ -479,7 +479,7 @@ struct RegTraits {
     computeSplitRegressionKernel<DataT, DataT, IdxT, TPB_DEFAULT>
       <<<grid, TPB_DEFAULT, smemSize, s>>>(
         b.pred, b.pred2, b.pred2P, b.pred_count, b.params.n_bins,
-        b.params.max_depth, b.params.min_samples_leaf, b.params.max_leaves,
+        b.params.max_depth, b.params.min_samples_split, b.params.max_leaves,
         b.input, b.curr_nodes, col, b.done_count, b.mutex, b.n_leaves, b.splits,
         b.block_sync, splitType);
   }