@hcho3 2020-11-18
The toy example used in the unit test was generated by the following Python script:
<details><summary>Very simple, naive implementation of decision tree fitting</summary>
<p>

Usage: `python test.py --max_depth=2 --min_samples_split=2 --min_samples_leaf=2`

```python
import numpy as np
import json
import argparse

class NumpyEncoder(json.JSONEncoder):
    def default(self, obj):
        if isinstance(obj, np.integer):
            return int(obj)
        elif isinstance(obj, np.floating):
            return float(obj)
        elif isinstance(obj, np.ndarray):
            return obj.tolist()
        return json.JSONEncoder.default(self, obj)

def mse(y):
    if y.shape[0] == 0:
        return 0
    y_pred = np.mean(y)
    return np.mean((y - y_pred)**2)

def initialize_tree(X, y):
    assert X.shape[0] == y.shape[0]
    return {'data': np.arange(X.shape[0]), 'leaf_value': np.mean(y)}

def find_best_split(X, y, node, *, min_samples_split, min_samples_leaf):
    ind = node['data']
    X_node, y_node = X[ind, :], y[ind]

    n_row, n_col = X_node.shape[0], X_node.shape[1]
    assert n_row == y_node.shape[0]
    if n_row < min_samples_split:
        return (None, None, 0.0)  # do not split
    best_gain = 0.0
    best_feature_id = None
    best_threshold = None
    for feature_id in range(n_col):
        feature_values = np.unique(X_node[:, feature_id])
        midpoint = (feature_values[:-1] + feature_values[1:]) / 2
        threshold_candidates = [-np.inf] + midpoint.tolist() + [np.inf]

        for threshold in threshold_candidates:
            left_ind = (X_node[:, feature_id] <= threshold)
            right_ind = (X_node[:, feature_id] > threshold)
            n_left, n_right = np.sum(left_ind), np.sum(right_ind)
            if n_left < min_samples_leaf or n_right < min_samples_leaf:
                continue
            y_left, y_right = y_node[left_ind], y_node[right_ind]
            assert n_left + n_right == n_row
            gain = mse(y_node) - n_left / n_row * mse(y_left) - n_right / n_row * mse(y_right)
            if gain > best_gain:
                best_gain = gain
                best_feature_id = feature_id
                best_threshold = threshold

    return (best_feature_id, best_threshold, best_gain)

def add_split(X, y, node, feature_id, threshold, gain):
    node['base_weight'] = node['leaf_value']
    del node['leaf_value']
    ind = node['data']
    X_node, y_node = X[ind, :], y[ind]
    left_ind = (X_node[:, feature_id] <= threshold)
    right_ind = (X_node[:, feature_id] > threshold)
    y_left, y_right = y_node[left_ind], y_node[right_ind]
    node['left'] = {'data': np.nonzero(left_ind)[0], 'leaf_value': np.mean(y_left)}
    node['right'] = {'data': np.nonzero(right_ind)[0], 'leaf_value': np.mean(y_right)}
    node['gain'] = gain
    node['feature_id'] = feature_id
    node['threshold'] = threshold

def fit(X, y, *, max_depth, min_samples_split, min_samples_leaf):
    tree = initialize_tree(X, y)
    queue = [tree]
    for i in range(max_depth):
        next_queue = []
        for node in queue:
            feature_id, threshold, gain = find_best_split(X, y, node,
                                                          min_samples_split=min_samples_split,
                                                          min_samples_leaf=min_samples_leaf)
            if gain > 0.0:
                add_split(X, y, node, feature_id, threshold, gain)
                next_queue.append(node['left'])
                next_queue.append(node['right'])
        queue = next_queue
        if not queue:
            break
    return tree

def main(args):
    X = np.array([[-1, 0], [0, 1], [2, 0], [0, 3], [-2, 0]], dtype=np.float32)
    y = X[:,0] + 2 * X[:,1]
    tree = fit(X, y, max_depth=args.max_depth, min_samples_split=args.min_samples_split,
               min_samples_leaf=args.min_samples_leaf)
    print(json.dumps(tree, indent=4, cls=NumpyEncoder))

if __name__ == '__main__':
    parser = argparse.ArgumentParser()
    parser.add_argument('--max_depth', type=int, default=2)
    parser.add_argument('--min_samples_split', type=int, default=0)
    parser.add_argument('--min_samples_leaf', type=int, default=0)
    args = parser.parse_args()
    main(args)

```

</p>
</details>