@hcho3 2020-11-18
> how does this change translate to the old backend?

The old backend was not changed, except for the renaming `min_rows_per_node` -> `min_samples_leaf`. The old backend won't support `min_samples_split`.