@hcho3 2020-11-18
The GPU test is failing with this error:

ImportError: /var/lib/jenkins/workspace/rapidsai/gpuci/cuml/prb/cuml-gpu-test/CUDA/10.2/OS/centos7/PYTHON/3.8/python/cuml/ensemble/randomforestclassifier.cpython-38-x86_64-linux-gnu.so: undefined symbol: _ZN2ML16set_rf_class_objEiifiiiifbbifiNS_9CRITERIONEbibi

The symbol represents
```
ML::set_rf_class_obj(int, int, float, int, int, int, int, float, bool, bool, int, float, int, ML::CRITERION, bool, int, bool, int)
```